requirements.txt
# Deep learning
torch>=2.1.0
torchvision>=0.16.0

# Core scientific
numpy>=1.24
scikit-learn>=1.3
scikit-image>=0.22

# Image & vision
opencv-python>=4.8
Pillow>=10.0
ImageHash>=4.3.1

# Utilities
tqdm>=4.66
pyyaml>=6.0
matplotlib>=3.8

# API
fastapi>=0.110
uvicorn>=0.27
python-multipart>=0.0.9

# Data
kagglehub
datasets>=2.18